Take the OSVersions related changes from the misclean ups.

+++ app/src/main/java/com/google/android/stardroid/touch/MapMover.java
@@ -51,7 +51,7 @@ public class MapMover implements
     this.model = model;
     this.controllerGroup = controllerGroup;
     Display display = ((WindowManager) context.getSystemService(
-        Context.WINDOW_SERVICE)).getDefaultDisplay(); 
+        Context.WINDOW_SERVICE)).getDefaultDisplay();
     int screenLongSize = display.getHeight();
     Log.i(TAG, "Screen height is " + screenLongSize + " pixels.");
     sizeTimesRadiansToDegrees = screenLongSize * Geometry.RADIANS_TO_DEGREES;
@@ -69,7 +69,7 @@ public class MapMover implements
   }
 
   @Override
-  public boolean onRotate(float degrees) {
+  public boolean onRotate(double degrees) {
     if (allowRotation) {
       controllerGroup.rotate(-degrees);
       return true;
@@ -79,14 +79,14 @@ public class MapMover implements
   }
 
   @Override
-  public boolean onStretch(float ratio) {
+  public boolean onStretch(double ratio) {
     controllerGroup.zoomBy(1.0f / ratio);
     return true;
   }
 
   @Override
   public void onSharedPreferenceChanged(SharedPreferences sharedPreferences,
-      String key) {
+                                        String key) {
     allowRotation = sharedPreferences.getBoolean(ALLOW_ROTATION, true);
   }
 }
